Daniel#C5
I worked on the help page. I moved the back button to the bottom and added a helpful tips message to the top. I also made it clear in the text what the question is and what the answer is, and I added a separator to make it look neater.